refactor(cli): remove TS backedges for shared paths
