+++ src/lib/http-probe.ts
@@ -11,11 +11,8 @@ import {
 } from "node:child_process";
 
 import type { ProbeResult } from "./onboard-types";
+import { ROOT } from "./paths";
 import { compactText } from "./url-utils";
-
-// runner.js is CJS — use require so we don't pull it into the TS build.
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const { ROOT } = require("../../bin/lib/runner");
 
 export type CurlProbeResult = ProbeResult;
