+++ src/lib/paths.ts
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import path from "node:path";
+
+export const ROOT = path.resolve(__dirname, "..", "..");
+export const SCRIPTS = path.join(ROOT, "scripts");
